requirements.txt
cprint==1.2.2
pandas==1.5.3
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
setuptools==65.6.3
streamlit==1.20.0
streamlit_lottie==0.0.3
